refactor to use req.product instead of always findById'ing closes #14
