+++ server/app/routes/index.js
@@ -2,7 +2,7 @@
 var router = require('express').Router();
 module.exports = router;
 
-router.use('/members', require('./members'));
+router.use('/users', require('./users'));
 router.use('/products', require('./products'));
 
 // Make sure this is after all of
